test: address final Base reliability review
