+++ packages/agent/test/finalization-recovery-sqlite-store.test.ts
@@ -1,8 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { spawn } from 'node:child_process';
-import { mkdtemp, readFile, rm, stat, symlink, writeFile } from 'node:fs/promises';
-import { existsSync } from 'node:fs';
-import { tmpdir } from 'node:os';
+import { readFile, rm, stat, symlink, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { DatabaseSync } from 'node:sqlite';
 import {
@@ -11,241 +8,15 @@ import {
 import {
   FINALIZATION_INBOX_DATABASE_FILENAME,
 } from '../src/finalization-recovery-store.js';
-import type {
-  VerifiedGraphScopedFinalizationEvidence,
-} from '../src/finalization-graph-envelope.js';
-
-const TX_HASH = `0x${'ab'.repeat(32)}`;
-const BLOCK_HASH = `0x${'cd'.repeat(32)}`;
-const RAW = Uint8Array.from([1, 2, 3, 4]);
-
-function received(overrides: Record<string, unknown> = {}) {
-  return {
-    key: 'entry-1',
-    chainId: 'base:84532',
-    contextGraphId: 'graph',
-    sourcePeerId: '12D3KooWPublisher',
-    ual: 'did:dkg:base:84532/0x1111111111111111111111111111111111111111/7',
-    txHash: TX_HASH,
-    assertionVersion: '1',
-    merkleRoot: `0x${'01'.repeat(32)}`,
-    kaId: '7',
-    batchId: '7',
-    targetContextGraphId: '42',
-    rawMessage: RAW,
-    ...overrides,
-  };
-}
-
-function evidence(
-  overrides: Partial<VerifiedGraphScopedFinalizationEvidence> = {},
-): VerifiedGraphScopedFinalizationEvidence {
-  return {
-    assertionVersion: '1',
-    publicTripleCount: 2,
-    privateTripleCount: 0,
-    publicQuadsDigest: `sha256:${'02'.repeat(32)}`,
-    publisherPeerId: '12D3KooWPublisher',
-    publisherAddress: '0x2222222222222222222222222222222222222222',
-    transactionHash: TX_HASH,
-    blockNumber: 123,
-    blockHash: BLOCK_HASH,
-    txIndex: 4,
-    authorAddress: '0x1111111111111111111111111111111111111111',
-    accessPolicy: 'ownerOnly',
-    allowedPeers: [],
-    ...overrides,
-  };
-}
-
-async function temporaryDirectory(): Promise<string> {
-  return mkdtemp(join(tmpdir(), 'dkg-finalization-inbox-'));
-}
-
-async function leaveCommittedReceivedInCrashedWal(path: string): Promise<void> {
-  const script = String.raw`
-const { createHash } = require('node:crypto');
-const { DatabaseSync } = require('node:sqlite');
-const path = process.env.DKG_FINALIZATION_INBOX_PATH;
-const raw = Buffer.from([1, 2, 3, 4]);
-const digest = createHash('sha256').update(raw).digest('hex');
-const database = new DatabaseSync(path);
-database.exec('PRAGMA journal_mode=WAL; PRAGMA synchronous=FULL; PRAGMA wal_autocheckpoint=0');
-database.prepare(
-  "INSERT INTO finalization_inbox_v1 (key,state,chain_id,context_graph_id,ual,tx_hash,"
-  + "assertion_version,merkle_root,ka_id,batch_id,envelope_sha256,raw_envelope,created_at,updated_at)"
-  + " VALUES (?,'RECEIVED',?,?,?,?,?,?,?,?,?,?,?,?)"
-).run(
-  'crash-entry', 'base:84532', 'graph',
-  'did:dkg:base:84532/0x1111111111111111111111111111111111111111/7',
-  '${TX_HASH}', '1', '0x${'01'.repeat(32)}', '7', '7', digest, raw, 1, 1
-);
-process.stdout.write('READY\n');
-setInterval(() => {}, 60_000);
-`;
-  const child = spawn(process.execPath, ['--experimental-sqlite', '-e', script], {
-    env: { ...process.env, DKG_FINALIZATION_INBOX_PATH: path },
-    stdio: ['ignore', 'pipe', 'pipe'],
-  });
-  let stderr = '';
-  child.stderr.setEncoding('utf8');
-  child.stderr.on('data', (chunk: string) => { stderr += chunk; });
-  await new Promise<void>((resolveReady, rejectReady) => {
-    let stdout = '';
-    const timeout = setTimeout(
-      () => rejectReady(new Error(`finalization inbox crash child timed out: ${stderr}`)),
-      10_000,
-    );
-    child.stdout.setEncoding('utf8');
-    child.stdout.on('data', (chunk: string) => {
-      stdout += chunk;
-      if (!stdout.includes('READY\n')) return;
-      clearTimeout(timeout);
-      resolveReady();
-    });
-    child.once('error', (error) => {
-      clearTimeout(timeout);
-      rejectReady(error);
-    });
-    child.once('exit', (code, signal) => {
-      clearTimeout(timeout);
-      rejectReady(new Error(
-        `finalization inbox crash child exited early: code=${code} signal=${signal} ${stderr}`,
-      ));
-    });
-  });
-  const exit = new Promise<void>((resolveExit) => child.once('exit', () => resolveExit()));
-  child.kill('SIGKILL');
-  await exit;
-}
-
-async function leaveCommittedV2MigrationInCrashedWal(path: string): Promise<void> {
-  const script = String.raw`
-const { DatabaseSync } = require('node:sqlite');
-const path = process.env.DKG_FINALIZATION_INBOX_PATH;
-const database = new DatabaseSync(path);
-database.exec('PRAGMA journal_mode=WAL; PRAGMA synchronous=FULL; PRAGMA wal_autocheckpoint=0');
-database.exec(
-  "BEGIN IMMEDIATE;"
-  + "CREATE TABLE finalization_pending_v2 ("
-  + "key TEXT PRIMARY KEY NOT NULL,chain_id TEXT NOT NULL,context_graph_id TEXT NOT NULL,"
-  + "source_peer_id TEXT,ual TEXT NOT NULL,tx_hash TEXT NOT NULL,assertion_version TEXT NOT NULL,"
-  + "merkle_root TEXT NOT NULL,ka_id TEXT NOT NULL,batch_id TEXT NOT NULL,"
-  + "target_context_graph_id TEXT,envelope_sha256 TEXT NOT NULL,raw_envelope BLOB NOT NULL,"
-  + "created_at INTEGER NOT NULL,updated_at INTEGER NOT NULL) STRICT;"
-  + "CREATE INDEX finalization_pending_time_v2 ON finalization_pending_v2(created_at,key);"
-  + "CREATE INDEX finalization_pending_graph_v2 ON finalization_pending_v2(context_graph_id,created_at);"
-  + "CREATE INDEX finalization_pending_peer_v2 ON finalization_pending_v2(source_peer_id,created_at);"
-  + "PRAGMA user_version=2;COMMIT;"
-);
-process.stdout.write('READY\n');
-setInterval(() => {}, 60_000);
-`;
-  const child = spawn(process.execPath, ['--experimental-sqlite', '-e', script], {
-    env: { ...process.env, DKG_FINALIZATION_INBOX_PATH: path },
-    stdio: ['ignore', 'pipe', 'pipe'],
-  });
-  let stderr = '';
-  child.stderr.setEncoding('utf8');
-  child.stderr.on('data', (chunk: string) => { stderr += chunk; });
-  await new Promise<void>((resolveReady, rejectReady) => {
-    let stdout = '';
-    const timeout = setTimeout(
-      () => rejectReady(new Error(`finalization inbox migration child timed out: ${stderr}`)),
-      10_000,
-    );
-    child.stdout.setEncoding('utf8');
-    child.stdout.on('data', (chunk: string) => {
-      stdout += chunk;
-      if (!stdout.includes('READY\n')) return;
-      clearTimeout(timeout);
-      resolveReady();
-    });
-    child.once('error', (error) => {
-      clearTimeout(timeout);
-      rejectReady(error);
-    });
-    child.once('exit', (code, signal) => {
-      clearTimeout(timeout);
-      rejectReady(new Error(
-        `finalization inbox migration child exited early: code=${code} signal=${signal} ${stderr}`,
-      ));
-    });
-  });
-  const exit = new Promise<void>((resolveExit) => child.once('exit', () => resolveExit()));
-  child.kill('SIGKILL');
-  await exit;
-}
+import {
+  BLOCK_HASH,
+  RAW,
+  evidence,
+  received,
+  temporaryDirectory,
+} from './finalization-recovery-sqlite-test-helpers.js';
 
 describe('SQLite finalization recovery store', () => {
-  it('migrates a v1 inbox in place without losing accepted entries', async () => {
-    const directory = await temporaryDirectory();
-    const databasePath = join(directory, FINALIZATION_INBOX_DATABASE_FILENAME);
-    try {
-      const initial = await openSqliteFinalizationRecoveryStore(directory);
-      await initial.receive(received());
-      await initial.close();
-
-      const legacy = new DatabaseSync(databasePath);
-      legacy.exec(`
-        DROP INDEX finalization_pending_peer_v2;
-        DROP INDEX finalization_pending_graph_v2;
-        DROP INDEX finalization_pending_time_v2;
-        DROP TABLE finalization_pending_v2;
-        PRAGMA user_version = 1;
-      `);
-      legacy.close();
-
-      const migrated = await openSqliteFinalizationRecoveryStore(directory);
-      expect(await migrated.list()).toMatchObject([{ key: 'entry-1', state: 'RECEIVED' }]);
-      const schema = new DatabaseSync(databasePath, { readOnly: true });
-      expect(schema.prepare('PRAGMA user_version').get()?.user_version).toBe(2);
-      expect(schema.prepare(
-        "SELECT name FROM sqlite_schema WHERE name = 'finalization_pending_v2'",
-      ).get()?.name).toBe('finalization_pending_v2');
-      schema.close();
-      await migrated.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('recovers a committed v2 migration when the main header still says v1', async () => {
-    const directory = await temporaryDirectory();
-    const databasePath = join(directory, FINALIZATION_INBOX_DATABASE_FILENAME);
-    try {
-      const initial = await openSqliteFinalizationRecoveryStore(directory);
-      await initial.receive(received());
-      await initial.close();
-
-      const legacy = new DatabaseSync(databasePath);
-      legacy.exec(`
-        DROP INDEX finalization_pending_peer_v2;
-        DROP INDEX finalization_pending_graph_v2;
-        DROP INDEX finalization_pending_time_v2;
-        DROP TABLE finalization_pending_v2;
-        PRAGMA user_version = 1;
-      `);
-      legacy.close();
-
-      await leaveCommittedV2MigrationInCrashedWal(databasePath);
-      expect(existsSync(`${databasePath}-wal`)).toBe(true);
-      expect((await readFile(databasePath)).readUInt32BE(60)).toBe(1);
-
-      const recovered = await openSqliteFinalizationRecoveryStore(directory);
-      expect(await recovered.list()).toMatchObject([{ key: 'entry-1', state: 'RECEIVED' }]);
-      const schema = new DatabaseSync(databasePath, { readOnly: true });
-      expect(schema.prepare('PRAGMA user_version').get()?.user_version).toBe(2);
-      expect(schema.prepare(
-        "SELECT name FROM sqlite_schema WHERE name = 'finalization_pending_v2'",
-      ).get()?.name).toBe('finalization_pending_v2');
-      schema.close();
-      await recovered.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
   it('lists only due live work in bounded oldest-first batches', async () => {
     const directory = await temporaryDirectory();
     try {
@@ -720,223 +491,6 @@ describe('SQLite finalization recovery store', () => {
     }
   });
 
-  it('durably defers a new envelope when live capacity is full', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, { maxEntries: 1 });
-      await store.receive(received());
-      await store.markVerified('entry-1', 0, evidence());
-      await expect(store.receive(received({
-        key: 'entry-2',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      expect(await store.list()).toMatchObject([{ key: 'entry-1', state: 'VERIFIED' }]);
-      expect(await store.health()).toMatchObject({
-        available: true,
-        ready: false,
-        degradedReason: 'capacity-exhausted',
-        deferredEntries: 1,
-      });
-
-      await store.transition('entry-1', 0, 'SETTLED');
-      await expect(store.promotePending(16)).resolves.toBe(1);
-      expect(await store.list()).toEqual(expect.arrayContaining([
-        expect.objectContaining({ key: 'entry-2', state: 'RECEIVED' }),
-      ]));
-      expect(await store.health()).toMatchObject({ deferredEntries: 0 });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('enforces the global deferred-entry quota independently', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 1,
-        maxDeferredEntries: 1,
-        maxDeferredPerPeer: 10,
-        maxDeferredPerContextGraph: 10,
-      });
-      await store.receive(received());
-      await expect(store.receive(received({
-        key: 'entry-2',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      await expect(store.receive(received({
-        key: 'entry-3',
-        contextGraphId: 'other-graph',
-        sourcePeerId: '12D3KooWOtherPublisher',
-        txHash: `0x${'12'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'capacity' });
-      expect(await store.health()).toMatchObject({ deferredEntries: 1 });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('enforces the per-peer deferred quota independently', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 1,
-        maxDeferredEntries: 10,
-        maxDeferredPerPeer: 1,
-        maxDeferredPerContextGraph: 10,
-      });
-      await store.receive(received());
-      await expect(store.receive(received({
-        key: 'entry-2',
-        contextGraphId: 'graph-2',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      await expect(store.receive(received({
-        key: 'entry-3',
-        contextGraphId: 'graph-3',
-        txHash: `0x${'12'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'capacity' });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('enforces the per-context-graph deferred quota independently', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 1,
-        maxDeferredEntries: 10,
-        maxDeferredPerPeer: 10,
-        maxDeferredPerContextGraph: 1,
-      });
-      await store.receive(received());
-      await expect(store.receive(received({
-        key: 'entry-2',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      await expect(store.receive(received({
-        key: 'entry-3',
-        sourcePeerId: '12D3KooWOtherPublisher',
-        txHash: `0x${'12'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'capacity' });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('enforces the deferred-byte quota independently', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 1,
-        maxDeferredEntries: 10,
-        maxDeferredBytes: 6,
-        maxDeferredPerPeer: 10,
-        maxDeferredPerContextGraph: 10,
-      });
-      await store.receive(received());
-      await expect(store.receive(received({
-        key: 'entry-2',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      await expect(store.receive(received({
-        key: 'entry-3',
-        contextGraphId: 'graph-3',
-        sourcePeerId: '12D3KooWOtherPublisher',
-        txHash: `0x${'12'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'capacity' });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('does not let a quota-blocked oldest graph starve a later eligible graph', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 2,
-        maxPerPeer: 8,
-        maxPerContextGraph: 1,
-      });
-      await store.receive(received({
-        key: 'live-graph-a',
-        contextGraphId: 'graph-a',
-        txHash: `0x${'a1'.repeat(32)}`,
-      }));
-      await store.receive(received({
-        key: 'live-graph-x',
-        contextGraphId: 'graph-x',
-        txHash: `0x${'a2'.repeat(32)}`,
-      }));
-      await expect(store.receive(received({
-        key: 'pending-graph-a',
-        contextGraphId: 'graph-a',
-        txHash: `0x${'b1'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-      await expect(store.receive(received({
-        key: 'pending-graph-b',
-        contextGraphId: 'graph-b',
-        txHash: `0x${'b2'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-
-      await store.transition('live-graph-x', 0, 'SUPERSEDED');
-      await expect(store.promotePending(1)).resolves.toBe(1);
-      expect(await store.get('pending-graph-b')).toMatchObject({
-        key: 'pending-graph-b',
-        state: 'RECEIVED',
-      });
-      expect(await store.get('pending-graph-a')).toBeUndefined();
-      expect(await store.health()).toMatchObject({ deferredEntries: 1 });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('accepts the 129th envelope durably and promotes it after the 128-entry backlog drains', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 128,
-        maxPerPeer: 128,
-        maxPerContextGraph: 128,
-      });
-      for (let index = 0; index < 128; index += 1) {
-        await expect(store.receive(received({
-          key: `entry-${index}`,
-          txHash: `0x${index.toString(16).padStart(64, '0')}`,
-          kaId: String(index),
-        }))).resolves.toMatchObject({ status: 'inserted' });
-      }
-      const overflow = received({
-        key: 'entry-128',
-        txHash: `0x${'ff'.repeat(32)}`,
-        kaId: '128',
-      });
-      await expect(store.receive(overflow)).resolves.toEqual({ status: 'pending' });
-      expect(await store.health()).toMatchObject({
-        deferredEntries: 1,
-        dueEntries: 128,
-      });
-
-      await store.transition('entry-0', 0, 'SUPERSEDED');
-      await expect(store.promotePending(16)).resolves.toBe(1);
-      expect(await store.get('entry-128')).toMatchObject({
-        state: 'RECEIVED',
-        kaId: '128',
-      });
-      expect(await store.health()).toMatchObject({ deferredEntries: 0 });
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
   it('serializes concurrent duplicate admission without duplicate rows', async () => {
     const directory = await temporaryDirectory();
     try {
@@ -970,28 +524,6 @@ describe('SQLite finalization recovery store', () => {
       const reopened = await openSqliteFinalizationRecoveryStore(directory);
       expect(await reopened.list()).toHaveLength(2);
       await reopened.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('recovers a committed RECEIVED row after a process crash with a hot WAL', async () => {
-    const directory = await temporaryDirectory();
-    try {
-      const initial = await openSqliteFinalizationRecoveryStore(directory);
-      const path = initial.databasePath;
-      await initial.close();
-
-      await leaveCommittedReceivedInCrashedWal(path);
-      expect(existsSync(`${path}-wal`)).toBe(true);
-
-      const recovered = await openSqliteFinalizationRecoveryStore(directory);
-      expect(await recovered.list()).toMatchObject([{
-        key: 'crash-entry',
-        state: 'RECEIVED',
-        rawMessage: RAW,
-      }]);
-      await recovered.close();
     } finally {
       await rm(directory, { recursive: true, force: true });
     }
@@ -1038,42 +570,6 @@ describe('SQLite finalization recovery store', () => {
         { key: 'verified', state: 'VERIFIED' },
         { key: 'trigger', state: 'RECEIVED' },
       ]);
-      await store.close();
-    } finally {
-      await rm(directory, { recursive: true, force: true });
-    }
-  });
-
-  it('expires stale deferred envelopes while preserving VERIFIED live evidence', async () => {
-    const directory = await temporaryDirectory();
-    let now = 1_000;
-    try {
-      const store = await openSqliteFinalizationRecoveryStore(directory, {
-        maxEntries: 1,
-        maxDeferredEntries: 1,
-        rawTtlMs: 100,
-        now: () => now,
-      });
-      await store.receive(received());
-      await store.markVerified('entry-1', 0, evidence());
-      await expect(store.receive(received({
-        key: 'stale-pending',
-        txHash: `0x${'ef'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-
-      now += 101;
-      await expect(store.receive(received({
-        key: 'fresh-pending',
-        contextGraphId: 'fresh-graph',
-        sourcePeerId: '12D3KooWFreshPublisher',
-        txHash: `0x${'12'.repeat(32)}`,
-      }))).resolves.toEqual({ status: 'pending' });
-
-      expect(await store.list()).toMatchObject([{ key: 'entry-1', state: 'VERIFIED' }]);
-      expect(await store.health()).toMatchObject({
-        deferredEntries: 1,
-        deferredPayloadBytes: RAW.byteLength,
-      });
       await store.close();
     } finally {
       await rm(directory, { recursive: true, force: true });
